index of midpointRays now have 6 parameters
We need to know where a midpointRay emits from. Therefore we introduced
following notation:
midpointRay_{x,y,z from x0,y0,z0}
where  x,y,z is the index of the region midpointRays midpoint lies on,
and x0,y0,z0 is the index of the region of the emitting midpoint.