New indices for midpointRays
We use some kind of combinatorial cube notation. The vertices of the
cube are denoted by "x,y,z" where x,y,z is either "+" or "-" depending
on if we increase the index from the emitting point of midpointRay to
the target region or if we decrease it. In the increase case we use -
else +