FIXED #3 negative y used
The bug was due to the insecurity with which GeoGebra assigns to
TPoint_1 the positive or the negative sign. In this case we simply had
to switch the sign in our definition of TPoint_{0,1,0} resp. -1. Since
the script is determinate and hopefully GeoGebra is, too, this should
fix the bug.