FIXED #8 embedded in html page
We can start the application in a browser as an html page and manipulate
the ggbapplet from the html then. See HelloWorld.html or
IS-projectionpoints,html for reference. It is important to note that we
first have to upload an empty GeoGebra worksheet which contains all our
neccessary tools to the tube, make it public and then can embed the
worksheet with its ID in our html.